# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/accounts/{id}":{"get":{"tags":["account-controller"],"operationId":"getAccountById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Account"}}}}}},"put":{"tags":["account-controller"],"operationId":"updateAccount","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}},"required":true},"responses":{"200":{"description":"OK"}}},"delete":{"tags":["account-controller"],"operationId":"deleteAccount","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/accounts/transferMoney/{fromAccountNumber}/{toAccountNumber}/{amount}":{"put":{"tags":["account-controller"],"operationId":"transferMoney","parameters":[{"name":"fromAccountNumber","in":"path","required":true,"schema":{"type":"string"}},{"name":"toAccountNumber","in":"path","required":true,"schema":{"type":"string"}},{"name":"amount","in":"path","required":true,"schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/accounts":{"get":{"tags":["account-controller"],"operationId":"getAllAccounts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Account"}}}}}}},"post":{"tags":["account-controller"],"operationId":"createAccount","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/accounts/byAccountNumber/{accountNumber}":{"get":{"tags":["account-controller"],"operationId":"getByAccountNumber","parameters":[{"name":"accountNumber","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"Account":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"accountNumber":{"type":"string"},"balance":{"type":"number","format":"double"}}}}}}